A      fix_apparmor_getprocattr.diff fix getprocattr so that it returns the start of the profile name string instead of the end of it.
A      rcu-locking-docs.txt
Change the locking docs to reflect the locking changes for rcu
aa_task_contexts

A      fix-rename-file-type.diff
Use the dentry->inode to determine if the medieated dentry is a directory.
The use of the inode that is pasted in is incorrect because it is
often the parent inode of the dentry being mediated.

A      misc-typos.diff
Fix a few miscellaneous typos

A      rcu-stale-forward-ptr.diff
A      rcu-caps-cache-comment.diff
Add a comment as to why resetting the caps cache instead of propogating it
is the right thing to do.

A      warn-info-messages.diff
Convert AA_WARN to us aa_audit_message, most of these messages while
not necessary for learning mode should be present in the logs with
the other apparmor messages.

A      rename-activate-unpack.diff
rename the interface routines to use unpack instead of activate, its
simpler and more standard.

j